renamed to ZosDsnList and removed full specifying of java.util.List in code and fixing merge conflicts
